fixed displaying dog details
