+++ webapp/src/dogDetails/dogDetails.test.tsx
@@ -5,18 +5,9 @@
 import React from "react";
 import { render, unmountComponentAtNode } from "react-dom";
 import { act, createRenderer } from "react-dom/test-utils";
-import DogDetails from "./dogDetails";
+import EditDetails from "./dogDetails";
 import { isNull } from "lodash";
 import { Button } from "@material-ui/core";
-import { store } from "../app/store";
-import { Provider } from "react-redux";
-import {
-  Redirect,
-  Route,
-  BrowserRouter as Router,
-  Switch,
-  useHistory,
-} from "react-router-dom";
 
 let container: HTMLDivElement | null = null;
 beforeEach(() => {
@@ -34,20 +25,19 @@ afterEach(() => {
   }
 });
 
-it("Rendered list of dogs", () => {
+it("Rendered register form button", () => {
   act(() => {
-    render(
-      <Provider store={store}>
-        <Router>
-        <Route path={"/"}
-            children={<DogDetails cookies={{}} dogId={1}/>}/>
-        </Router>
-      </Provider>,
-      container
-    );
+    render(<EditDetails />, container);
   });
   expect(container?.getElementsByClassName("mainForm").length).toEqual(1);
   expect(
     container?.getElementsByClassName("classes.formControl").length
   ).toBeGreaterThanOrEqual(0);
+});
+
+it("test only register button", () => {
+  act(() => {
+    render(<EditDetails />, container);
+  });
+  const myRegisterClicked = jest.fn();
 });
