+++ webapp/src/editDogDetails/editDogDetails.test.tsx
@@ -5,9 +5,18 @@
 import React from "react";
 import { render, unmountComponentAtNode } from "react-dom";
 import { act, createRenderer } from "react-dom/test-utils";
-import EditDetails from "./editDetails";
+import EditDogDetails from "./editDogDetails";
 import { isNull } from "lodash";
 import { Button } from "@material-ui/core";
+import { store } from "../app/store";
+import { Provider } from "react-redux";
+import {
+  Redirect,
+  Route,
+  BrowserRouter as Router,
+  Switch,
+  useHistory,
+} from "react-router-dom";
 
 let container: HTMLDivElement | null = null;
 beforeEach(() => {
@@ -25,19 +34,20 @@ afterEach(() => {
   }
 });
 
-it("Rendered register form button", () => {
+it("Rendered list of dogs", () => {
   act(() => {
-    render(<EditDetails />, container);
+    render(
+      <Provider store={store}>
+        <Router>
+        <Route path={"/"}
+            children={<EditDogDetails cookies={{}} dogId={1}/>}/>
+        </Router>
+      </Provider>,
+      container
+    );
   });
   expect(container?.getElementsByClassName("mainForm").length).toEqual(1);
   expect(
     container?.getElementsByClassName("classes.formControl").length
   ).toBeGreaterThanOrEqual(0);
-});
-
-it("test only register button", () => {
-  act(() => {
-    render(<EditDetails />, container);
-  });
-  const myRegisterClicked = jest.fn();
 });
